Se completó el documento de modelos de mezclas, no se pusieron todas las implementaciones debido a la alta complejidad del problema, pero se agregaron ejemplos, casos de estudio y muchas referencias.

+++ clasificacion/no-supervisado/kMeans.py
@@ -89,7 +89,7 @@ class kMeans:
                 self.centroides[k] = np.mean(v, axis=0)
 
         # Crear arreglo de etiquetas
-        y_etiquetas = np.zeros(datos.shape[0])
+        y_etiquetas = np.zeros(datos.shape[0], dtype=int)
         for k, v in self.clusters.items():
             for dato in v:
                 idx_dim, _ = np.where(datos == dato)
